refactor: ban user modal ui
