+++ src/features/ban-user/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './BanUserModal'
